adding validation checks

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const { page: currentPage } = req.query || 1;
@@ -86,21 +87,28 @@ exports.createPost = (req, res, next) => {
     // imageUrl = req.file.path.replace("\\","/");
     const imageUrl = req.file.path;
     const { title, content } = req.body;
+    let creator;
     const post = new Post({
         title,
         content,
         imageUrl,
-        creator: {
-            name: 'Conary'
-        },
+        creator: req.userId,
     });
 
     post.save()
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully',
-                post: result
+                post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -148,6 +156,14 @@ exports.updatePost = (req, res, next) => {
                 throw error;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.');
+
+                error.statusCode = 403;
+
+                throw error;
+            }
+
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -187,18 +203,33 @@ exports.deletePost = (req, res, next) => {
 
                 throw error;
             }
-            // check user
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.');
+
+                error.statusCode = 403;
+
+                throw error;
+            }
             clearImage(post.imageUrl);
 
             return Post.findByIdAndRemove(postId);
         })
-        .then((result) => {
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+
+            return user.save();
+        })
+        .then(() => {
             res
                 .status(200)
                 .json({
                     message: 'Post Deleted'
                 });
-        }).catch(err => {
+        })
+        .catch(err => {
             if (!err.statusCode) {
                 err.statusCode = 500;
             }
